Add `text.json.ParserMarker`, a type that can be thrown into a decoded struct to match any JSON value.
As opposed to `JSONValue`, `ParserMarker` just skips the region in the input stream without parsing. Later on, you can call `marker.decode!(T, handler)` to resume parsing inside the marker.

Note that while faster than parsing, skipping in the input stream is not free! The JSON will still need to be lexed. Only use `ParserMarker` if there is no way to know the full type of the message at the time.